Implement conversion to snake_case/camelCase strategy and consolidate changes into one file
